Frontend service refactor (#124)
