+++ blotztask-ui/src/services/taskService.ts
@@ -2,7 +2,7 @@ import { TaskItemDTO } from "@/model/task-Item-dto";
 import { fetchWithAuth } from "@/utils/fetch-with-auth";
 
 export const fetchAllTaskItems = async (): Promise<TaskItemDTO[]> => {
-  const response = await fetchWithAuth(
+  const result = await fetchWithAuth<TaskItemDTO[]>(
     `${process.env.NEXT_PUBLIC_API_BASE_URL_WITH_API}/Task/alltask`,
     {
       method: 'GET',
@@ -12,10 +12,5 @@ export const fetchAllTaskItems = async (): Promise<TaskItemDTO[]> => {
     }
   );
 
-  if (!response.ok) {
-    throw new Error('Network response was not ok');
-  }
-
-  const data: TaskItemDTO[] = await response.json();
-  return data;
+  return result;
 };
