+++ blotztask-ui/src/utils/fetch-with-auth.ts
@@ -1,6 +1,7 @@
 import Cookies from "js-cookie"; // Use js-cookie for client-side cookies
+import { fetchWithErrorHandling } from "./http-client";
 
-export const fetchWithAuth = async (url: string, options: RequestInit = {}) => {
+export const fetchWithAuth = async<T>(url: string, options: RequestInit = {}) => {
   // Get the auth token from client-side cookies
   const accessToken = Cookies.get('authToken');
 
@@ -15,7 +16,7 @@ export const fetchWithAuth = async (url: string, options: RequestInit = {}) => {
   };
 
   // Return the fetch call with the attached headers
-  return fetch(url, {
+  return fetchWithErrorHandling<T>(url, {
     ...options,
     headers: authHeaders,
   });
